edited robot dimensions

+++ src/diffdrive_arduino/include/diffdrive_arduino/wheel.hpp
@@ -18,13 +18,13 @@ class Wheel
 
     Wheel() = default;
 
-    Wheel(const std::string &wheel_name, int counts_per_rev)
+    Wheel(const std::string &wheel_name, double counts_per_rev)
     {
       setup(wheel_name, counts_per_rev);
     }
 
     
-    void setup(const std::string &wheel_name, int counts_per_rev)
+    void setup(const std::string &wheel_name, double counts_per_rev)
     {
       name = wheel_name;
       rads_per_count = (2*M_PI)/counts_per_rev;
